requirements.txt
gdeltdoc>=0.2.0
pymongo>=4.6.0
python-dotenv>=1.0.0
feedparser>=6.0.0
